Hope to fix the NAND-flash blocking on nosmp
There is a strange problem on the driver, it can block the NAND-Flash
of the Easybox 904x see:
https://github.com/Quallenauge/lede-feeds-easybox904/commit/
1fea28f8156344a6f7609f3e4becc0b969943d93

It was here optimized for SMP use, but on nosmp use (with telefon support)
Bigger files that are exist from the beginning can not be change,
because the Router hangs.

i mix the new one with the old, and do only short tests on nosmp build.

Signed-off-by: Jonas Albrecht <plonkbong100@protonmail.com>